Update state dict and model together (#573)
* code beautification

* code beautification, move functions together

* rewrite model rewriter

* rewrite quantizers

* weights is none check

* typo

* not weight -> weight is not None

* fix dimensions for parallel prefill

* test

* typo

* bfloat16 on ARM with MacOS 14

* precision for a8w4

* sdpa_kv

* fixes

* inline qlq definition

* trial and error

* qdq not working

* ci

* not so fast with bf16=fast

* typo, and handle fast across maxcos version...

* typo

* type cast

+++ export_et_util.py
@@ -63,15 +63,16 @@ class CustomSDPAAttention(nn.Module):
         k = k.view(bsz, seqlen, self.n_local_heads, self.head_dim)
         v = v.view(bsz, seqlen, self.n_local_heads, self.head_dim)
 
-        q = apply_rotary_emb(q, freqs_cis)
-        k = apply_rotary_emb(k, freqs_cis)
-
+        q = apply_rotary_emb(q, freqs_cis).to(dtype=torch.float)
+        k = apply_rotary_emb(k, freqs_cis).to(dtype=torch.float)
+        v = v.to(dtype=torch.float)
+        
         # KV cache should always be enabled
         assert self.kv_cache is not None
         output = torch.ops.llama.sdpa_with_kv_cache(
-            q.float(),
-            k.float(),
-            v.float(),
+            q,
+            k,
+            v,
             self.kv_cache.k_cache,
             self.kv_cache.v_cache,
             input_pos[-1].item(),
